refactor: 스케줄러 -> 배치로 변경
