+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'budong-scheduler'
+rootProject.name = 'budong-batch'
