+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=budong-scheduler
+spring.application.name=budong-batch
